code/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
flake8==7.2.0
fonttools==4.57.0
fpdf2==2.8.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
identify==2.6.10
idna==3.10
jiter==0.9.0
mccabe==0.7.0
nodeenv==1.9.1
openai==1.75.0
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.51
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
smmap==5.0.2
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.30.0
wcwidth==0.2.13
